+++ packages/button/src/LionButton.js
@@ -1,8 +1,13 @@
-import { css, html, SlotMixin, DisabledWithTabIndexMixin, LitElement } from '@lion/core';
+import {
+  css,
+  html,
+  browserDetection,
+  SlotMixin,
+  DisabledWithTabIndexMixin,
+  LitElement,
+} from '@lion/core';
 
-// eslint-disable-next-line class-methods-use-this
 const isKeyboardClickEvent = e => e.keyCode === 32 /* space */ || e.keyCode === 13; /* enter */
-// eslint-disable-next-line class-methods-use-this
 const isSpaceKeyboardClickEvent = e => e.keyCode === 32; /* space */
 
 export class LionButton extends DisabledWithTabIndexMixin(SlotMixin(LitElement)) {
@@ -27,7 +32,7 @@ export class LionButton extends DisabledWithTabIndexMixin(SlotMixin(LitElement))
     return html`
       <div class="btn">
         ${this._renderBefore()}
-        ${this.constructor.__isIE11()
+        ${browserDetection.isIE11
           ? html`
               <div id="${this._buttonId}"><slot></slot></div>
             `
@@ -153,11 +158,11 @@ export class LionButton extends DisabledWithTabIndexMixin(SlotMixin(LitElement))
     this.active = false;
     this.__setupDelegationInConstructor();
 
-    if (this.constructor.__isIE11()) {
+    if (browserDetection.isIE11) {
       this._buttonId = `button-${Math.random()
         .toString(36)
         .substr(2, 10)}`;
-      this.setAttribute('aria-labelledby', this._buttonId);
+      this.updateComplete.then(() => this.setAttribute('aria-labelledby', this._buttonId));
     }
   }
 
@@ -254,11 +259,5 @@ export class LionButton extends DisabledWithTabIndexMixin(SlotMixin(LitElement))
       // dispatch click
       this.click();
     }
-  }
-
-  static __isIE11() {
-    const ua = window.navigator.userAgent;
-    const result = /Trident/.test(ua);
-    return result;
   }
 }