+++ packages/choice-input/index.js
@@ -1,1 +1,2 @@
+export { ChoiceGroupMixin } from './src/ChoiceGroupMixin.js';
 export { ChoiceInputMixin } from './src/ChoiceInputMixin.js';