+++ packages/checkbox/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.11](https://github.com/ing-bank/lion/compare/@lion/checkbox@0.3.10...@lion/checkbox@0.3.11) (2020-02-10)
+
+**Note:** Version bump only for package @lion/checkbox
+
+
+
+
+
+## [0.3.10](https://github.com/ing-bank/lion/compare/@lion/checkbox@0.3.9...@lion/checkbox@0.3.10) (2020-02-06)
+
+**Note:** Version bump only for package @lion/checkbox
+
+
+
+
+
+## [0.3.9](https://github.com/ing-bank/lion/compare/@lion/checkbox@0.3.8...@lion/checkbox@0.3.9) (2020-02-06)
+
+**Note:** Version bump only for package @lion/checkbox
+
+
+
+
+
+## [0.3.8](https://github.com/ing-bank/lion/compare/@lion/checkbox@0.3.7...@lion/checkbox@0.3.8) (2020-02-06)
+
+**Note:** Version bump only for package @lion/checkbox
+
+
+
+
+
 ## [0.3.7](https://github.com/ing-bank/lion/compare/@lion/checkbox@0.3.6...@lion/checkbox@0.3.7) (2020-02-05)
 
 **Note:** Version bump only for package @lion/checkbox