+++ packages/helpers/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1](https://github.com/ing-bank/lion/compare/@lion/helpers@0.3.0...@lion/helpers@0.3.1) (2020-02-06)
+
+**Note:** Version bump only for package @lion/helpers
+
+
+
+
+
 # [0.3.0](https://github.com/ing-bank/lion/compare/@lion/helpers@0.2.2...@lion/helpers@0.3.0) (2020-01-21)