+++ packages/localize/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.6](https://github.com/ing-bank/lion/compare/@lion/localize@0.8.5...@lion/localize@0.8.6) (2020-02-06)
+
+**Note:** Version bump only for package @lion/localize
+
+
+
+
+
 ## [0.8.5](https://github.com/ing-bank/lion/compare/@lion/localize@0.8.4...@lion/localize@0.8.5) (2020-02-03)