+++ packages/switch/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.13](https://github.com/ing-bank/lion/compare/@lion/switch@0.5.12...@lion/switch@0.5.13) (2020-02-10)
+
+**Note:** Version bump only for package @lion/switch
+
+
+
+
+
+## [0.5.12](https://github.com/ing-bank/lion/compare/@lion/switch@0.5.11...@lion/switch@0.5.12) (2020-02-06)
+
+**Note:** Version bump only for package @lion/switch
+
+
+
+
+
+## [0.5.11](https://github.com/ing-bank/lion/compare/@lion/switch@0.5.10...@lion/switch@0.5.11) (2020-02-06)
+
+**Note:** Version bump only for package @lion/switch
+
+
+
+
+
+## [0.5.10](https://github.com/ing-bank/lion/compare/@lion/switch@0.5.9...@lion/switch@0.5.10) (2020-02-06)
+
+**Note:** Version bump only for package @lion/switch
+
+
+
+
+
 ## [0.5.9](https://github.com/ing-bank/lion/compare/@lion/switch@0.5.8...@lion/switch@0.5.9) (2020-02-05)
 
 **Note:** Version bump only for package @lion/switch